requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-bcrypt
psycopg2-binary
anthropic
python-dotenv